Rename getFileTreeStatus -> lStatTree.
getFileTreeStatus looks like an extension of getFileStatus, but it's
actually more like getSymbolicLinkStatus. getSymbolicLinkStatus is just
too long, but let's give it a name that more accurately hints at it's
semantics.

The prior name isn't coincidental -- when it was named, it *was* an
extension of getFileStatus, but that turned out to be not what we want.